packages/eoc.core/pyproject.toml
[project]
name = "eoc.core"
version = "0.1.0"
description = "A library setting the foundation for EOC (Era of Challenges) infrastructure, based on the `warelib` library."
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.2",
    "warelib @ git+https://github.com/pygame-community/warelib.git@main",
    "pygame-gui>=0.6.13",
]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/eoc"]
